feat: Implement restaurant ownership and authorization features
- Added DeleteDishesForRestaurant command and handler for removing all dishes from a restaurant.
- Enhanced CreateRestaurant command handler to associate the restaurant with the current user as the owner.
- Updated DeleteRestaurant command handler to include authorization checks for deleting a restaurant.
- Introduced UpdateRestaurant command and handler with authorization checks for updating restaurant details.
- Created AssignUserRole and UnassignUserRole commands and handlers for managing user roles.
- Removed CurrentUser and UserContext classes, replacing them with a more streamlined approach for user management.
- Added ResourceOperation enum and IRestaurantAuthorizationService interface for better authorization handling.
- Implemented RestaurantAuthorizationService to manage authorization logic based on user roles and ownership.
- Updated database schema to include OwnerId in the Restaurants table and added necessary migrations.
- Seeded initial user roles (User, Owner, Admin) into the database.

+++ Restaurants.Application/Restaurants/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Authorization.Enums;
+using Restaurants.Domain.Authorization.Interfaces;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
@@ -7,7 +9,8 @@ using Restaurants.Domain.Repositories;
 namespace Restaurants.Application.Restaurants.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-                                            IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+                                            IRestaurantsRepository restaurantsRepository,
+                                            IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -16,9 +19,10 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
         var restaurant = await restaurantsRepository.GetByIdAsync(request.Id);
 
         if (restaurant is null)
-        {
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
-        }
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbiddenException();
 
         await restaurantsRepository.DeleteAsync(restaurant);
     }
